refactor: extract the argv and path guard helpers into argv.py
Moves 36 top-level symbols (18 functions/helpers + their length/shape
constants and regexes) from server.py into a new leaf module,
src/comfy_mcp/argv.py, byte-for-byte: the shared argument-injection and
OS-limit primitives (_reject_nul, _reject_option_like, _reject_nul_deep,
_guard_arg_len, _encode_argv, _bounded_timeout, _clip_for_error) and
every per-domain guard built on them (_guard_workflow_path,
_guard_prompt_id, _guard_download_id, _guard_extra_args, _guard_version,
_guard_node_names, _guard_log_port/_render_bad_port,
_guard_model_relative_path/_guard_model_filename,
_validate_upload_paths).

Symbols were re-located by AST, not by the plan's (now stale) line
numbers, and a handful of length constants the plan's prose didn't
name individually (_MAX_EXTRA_ARG_LEN, _MAX_NODE_PACK_NAMES/_CHARS,
_REGISTRY_ID_RE, _MIN_LOG_PORT/_MAX_LOG_PORT/_MAX_PORT_REPR_CHARS) moved
too, once an AST call-graph scan showed each is read, in code, only by
a moving guard — leaving them behind would have put a back-edge from
argv into server. Two of the plan's named clusters are non-contiguous
in the current file: _guard_node_names's three supporting constants
sit ~150 lines ahead of the function, separated by node-install
consent machinery that stays in server.py; _MAX_UPLOAD_PATHS(_TOTAL_BYTES)
sit next to _UPLOAD_STDOUT_MAX_CHARS, which stays (it sizes a tool
that isn't moving). Verified byte-for-byte via ast.get_source_segment
against pre-move HEAD.

server.py reaches every moved name module-qualified (argv._x) at every
call site and prose cross-reference — 146 sites, via a word-boundary
rename plus the ruff F821 net, which came back clean (zero missed call
sites). ntpath dropped from server.py's imports (its only user moved).
One real bug surfaced by the full test run rather than ruff: argv is
also a pre-existing local variable name in _generate_param_args and
_run_template_param_args (the built --name=value / --param=KEY=VALUE
token list), so the blanket rename silently shadowed the module inside
those two functions. Fixed by renaming the locals to param_args
(already the name a sibling function's parameter uses for the same
value) rather than the module.

Eleven test files rename 81 read-only references plus the one
monkeypatch (tests/test_workflow.py's _MAX_PATH_ARG_LEN patch moves
from server to argv, docstring intent unchanged); conftest.py and
test_spawn_failure.py keep their argv._x docstring mentions unqualified
by an import, matching the existing clitext._x-without-import
precedent. AGENTS.md's module table gains an argv.py row ("Six leaf
modules" -> "Seven").

Full suite: 1889 passed, 3 deselected. ruff check and ruff format
--check both clean.
